+++ test/test_filelist.py
@@ -101,7 +101,7 @@ def test_filelist(testdirectory):
         os.path.join(testdirectory.path(), 'a', '*')
     ]
 
-    filelist = giit.filelist.FileList(
+    filelist = giit.filelist.FileMapper(
         local_path=testdirectory.path(),
         remote_path=os.path.join(os.path.sep, 'var', 'www'),
         exclude_patterns=excludes)
@@ -147,7 +147,7 @@ def test_filelist_excludes_config(testdirectory):
         testdirectory.path() + '/a/*'
     ]
 
-    filelist = giit.filelist.FileList(
+    filelist = giit.filelist.FileMapper(
         local_path=testdirectory.path(),
         remote_path=os.path.join(os.path.sep, 'var', 'www'),
         exclude_patterns=excludes)
@@ -180,3 +180,18 @@ def test_filelist_excludes_config(testdirectory):
     assert file2 in remote_result
     assert file3 in remote_result
     assert file4 in remote_result
+
+
+def test_filelist(testdirectory):
+    a_dir = testdirectory.mkdir('a')
+    a_dir.write_text(filename='a.txt', data=u'a', encoding='utf-8')
+
+    filelist = giit.filelist.FileList(
+        from_path=testdirectory.path(),
+        exclude_patterns=[])
+
+    result = list(filelist)
+
+    assert len(result) == 1
+
+    assert os.path.join('a', 'a.txt') == result[0]