+++ NEWS.rst
@@ -9,6 +9,8 @@ Latest
 * Minor: Better support build building local branches (which does
   not yet exist on the remote).
 * Minor: Clone from path if available.
+* Minor: Support url_base for push type steps. This will print
+  all index.html files pushed.
 
 2.0.0
 -----