+++ NEWS.rst
@@ -6,7 +6,11 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Better support build building local branches (which does
+  not yet exist on the remote).
+* Minor: Clone from path if available.
+* Minor: Support publish_url for push type steps. This will print
+  all index.html files pushed.
 
 2.0.0
 -----