+++ NEWS.rst
@@ -6,6 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
+* tbd
+
+2.1.0
+-----
 * Minor: Better support build building local branches (which does
   not yet exist on the remote).
 * Minor: Clone from path if available.